ARM: pxa: power: enable console suspend
Remove "no_console_suspend" in kernel command line.

After removing this, console will be suspended when entering suspend.
So part of the suspend log will only be displayed when resume,
including device suspend and device late suspend.

This patch is also a WR of a console issue. Before this patch, after
several suspend/resume sequence, console may be locked if issue
several characters at one time. This issue is not root caused but
can be work around by this patch.

Change-Id: I8ba8cfa11d49ce0b2d67b8abc4275492c701a320
Signed-off-by: Lu Mao <ylmao@marvell.com>