ARM: mmp: hook cpu notify for hotpulg
This is based on the commit "634dda9" and enhance for the enclusive
operations with hotplug, DFC and C2.

The deadlock which described in the commit "634dda9" is not only exist
b/t hotplug and C2, also it will exist b/t hotplug and DFC. So the more
reasonable method is to hook the cpu notify for hotplug; when the
action is preparing cpu down/up, then need set the flag to disable
DFC/C2; after receive the notify for the operations done, we need enable
DFC/C2 again.

Change-Id: I3fa5a21f9334cd3bbc4b1f2cd492aaba663ee2ed
Signed-off-by: Leo Yan <leoy@marvell.com>