ARM: mmp: CPUFreq: remove highest pp limitation
If set cpumask for the policy, then it will only set the cpu frequency
to core 0, so this will introduce a bug by the highest pp limitation:
when core 1 is to be pluged in, it will set the highest pp limitation
in the init function; so it will set the highest pp limitation at this
point, sometimes it will run at 1GHz and set this value at the
limitation; so later the cpufreq governor cannot set lower frequency by
this limitation.

This patch will remove the highest pp limitation, for now we have set all
cpu mask for the policy, so that they have bound.

Change-Id: I923aa1121876d3cef77267f7c17f1352d6c6eea9
Signed-off-by: Leo Yan <leoy@marvell.com>