mmc: sdhci: keep the saved clock var up to date
the clock rate set to the sd controller may not exactly as requested
by the mmc core, this patch make the clock rate saved in the mmc_host
and sdhci_host updated with the real setting as in the controller. Thus
"/sys/kernel/debug/mmc0/ios" and card detect prints can show the correct
clock rate.

this patch also fixes a typo from mhz to Hz of the clock rate print.

Change-Id: Ib920f3f8b5418d3875bc1f308ed4f8f3f1154b82
Signed-off-by: Bin Wang <binw@marvell.com>