usb:diag: fix logic bug in pxa910_gs_write
When connect the usb to a PC that does not have ACAT driver, the buffer will
be full in a short time, and the function will always return -EAGAIN in the
middle of the function. When re-connect the USB to a pc with ACAT driver,
the buffer is still full, and there is no chance to trigger the USB to start
sending message by calling pxa910_gs_start_tx(port).
Move the error handling code to the end of the function can fix this bug.

Change-Id: I5d0c859eb044844220f460039266292a7c35503b
Signed-off-by: Leo Song <liangs@marvell.com>