ARM: gic: add irq_set_affinity to gic_arch_extn
Some architectures would shut down the mp logic for certain low power
mode cases, thus it relies on an external interrupt controller to
wake up the cores. In order to keep the same interrupt distribute
mapping as in the gic, here export the irq_set_affinity hook in the
gic_arch_extn.

Change-Id: I8c70b03ba34df7370739f27a806207ba37f5c560
Signed-off-by: Bin Wang <binw@marvell.com>