soc-camera: change default format strategy
In previous implementation, soc_camera will apply a default format
to sensor as soon as the video device is opened. This result in a
long time to open the video device, a lot of i2c transaction done
during the open time.

Change strategy to check if controller format and sensor format
consist just beform streamon, and only apply to sensor if they are
not consist.

Change-Id: Iad2e9d8af19729de592c392860e1e4a551cfe289
Signed-off-by: Jiaquan Su <jqsu@marvell.com>