ARM: mmp: refine c2 exit flow
When exit from c2, if the core is waken up by irqs then let core wait
for the semaphore.

In previous code it will try to acquire the semaphore, if it's failure
then it will jump back to c2 again. now just let the core wait for
the semaphore until it can acquire it successfully.

Change-Id: I997cc820a4cd144a997711f2bb2f57b95f93b40e
Signed-off-by: Leo Yan <leoy@marvell.com>