pxa95x: fix pm overheating init function
This Patch fix a kernel panic in pxa95x_pm.c

The issue was that when board boots up with high temperture
the overheating irq will be triggered as soon as the
request irq is called.
in this case workque and timers are not inited.

This patch revers the init order to make sure workque and timer
are created befor the call to request_irq.

Change-Id: I3fdad7219f0ca0c2c29511fb94ef821513344b52
Signed-off-by: Chen Reichbach <creichba@marvell.com>