ARM: pxa: handle every pending interrupt in gpio demux handler
The gpio demux handler will read pending interrupt status in GEDR
register and leave only those unmasked interrupts for further
processing. This will introduce a race condition if (1) the
irq_disable function in pxa_muxed_gpio_chip is implemented as the
irq_mask function and (2) an interrupt happens when you disable_irq()
in your driver.

In disable_irq() it will call irq_chip's irq_disable() function if
it is defined. If irq_disable() is implemented like irq_mask(), then
the irq will be masked and its status will be recorded in irq_mask
flag. If an interrupt happens at that time and when disable_irq()
returns, this interrupt will trigger pxa_gpio_demux_handler(). As
it is masked in irq_mask, it will not be cleared in GEDR register
and the following generic_handle_irq() will not be executed. When
it returns, the same interrupt will trigger pxa_gpio_demux_handler()
again. Thus we'll meet irq storms in this case.

The correct way to handle it is to bypass irq_mask check in gpio
demux handler and let irq's handler (handle_edge_irq() in this case)
to deal with it directly. If the irq is disabled, handle_edge_irq()
will mark it pending, mask_ack_irq and quit.

Also note that as handle_edge_irq() will ack the irq anyway we need
not do it in pxa_gpio_demux_handler() either.

Just remove the irq_mask check and push the irq ack operation to
irq's flow handler to make the gpio interrupt handling clearer.

Change-Id: I48d2fb0de1f679dda8ce0615bf9b0f82f8e4c263
Signed-off-by: Frank Zhang <zyifan@marvell.com>
Signed-off-by: Fan Wu <fwu@marvell.com>
Signed-off-by: Ning Jiang <ning.jiang@marvell.com>