uart: add flag to prevent uart suspend
This patch adds a flag to prevenet uart suspend.
This flag is passed from platform info and makes
uart suspend callback function not called so in
fake suspend uart would be alive
This patch fixes the bug that after entering suspend
for the first time, we can't copy&paste commands
to terminal that are more than 32 bytes.

Change-Id: Iebea664f3faddf01c1c8268897f20c36f95e33e8
Signed-off-by: Moran Raviv <moranr@marvell.com>