ARM: pxa978: move TEXT_BASE to 0x01088000
Move TEXT_BASE to 0x01088000 to reserve additional
512KB space for future usage(such as ram console and emmd).

We cannot move kernel more because of module branch limitation
Currently, modules are loaded to VA (0xc0000000 - 8MB).
Modules to kernel branch should within 32MB range,
so kernel code should be loaded before 0xc0000000 + 24MB.

On nevo, VA 0xc0000000 is mapped to PA 0x80000000 and
the first 16MB is reserved to CP. So all of kernel code
should be located between 0x81000000 to 0x81800000.
This is the reason why we cannot push kernel to high memory
address and we also should keep kernel size as small as possible

Change-Id: I7613ab840caa9ca73d2ddb82aadf498551f72379
Signed-off-by: Xiaofan Tian <tianxf@marvell.com>