ARM: pxa: Use AVLSR instead of previous AVLCR to get current voltage level
Previously we use AVLCR to get voltage level which is wrong.

The definition of the AVLCR
AVLCR : Application Subsystem Voltage Level Change Register.

There is a Register to describe the current voltage status, which is AVLSR.
Definition as follows:
AVLSR : Application Subsystem Voltage Level Status Register

AVLCR is used to change voltage level while AVLSR is used to get voltage level.

Thus, replace AVLCR with AVLSR when we need to get the voltage level.

Change-Id: I63b2dc57eae7a10db35efda47ad0b07efc5e748f
Signed-off-by: fwu <fwu@marvell.com>