ARM: pxa: power: keep ABU clock in D0CG/D0CG-LCD on Nevo D0
For Nevo D0, ABU can be still working in D0CG/D0CG-LCD. So we can't
rudely turn off its clock in D0CG/D0CG-LCD entry.

This patch is to keep the two ABU clocks, CKENA[ABU SSI] and
CKENB[ABU] no change in the entry of D0CG/D0CG-LCD. And the on/off
is decided by ABU driver.

This patch also changed the wakeup event of D0CG/D0CG-LCD. ABU have
to be a wakeup event.

Change-Id: I6ef9c9c1d741dd78dc0ed50f5df3ebe9016827d5
Signed-off-by: Lu Mao <ylmao@marvell.com>