ARM: pxa: power: prolong the VCTCXO stablization time to 2ms for DKB
To make Comm side more stble, we prolong the VCTCXO stablization time
to ~2ms from ~0.5ms set before. But this setting will make D1/D2 exit
time longer.

From test, we found that the exit time is about 1.7ms longer.
And the MP3 scenario power increases about 1mA.

So this patch is just for system stablization and this parameter need
to be tuned later for the balance between stablization and power.

Change-Id: Ibd72bff01bb94666e37dcc6a38699bea83a99ebb
Signed-off-by: Lu Mao <ylmao@marvell.com>