pxa95xfb: remove eof_intr_en flag in fb
this flag is used to select fetch eof on/off status but it's not
required now. fetch eof would be on if eof_handler function pointer
is used.
remove this flag would:
1. make code more clear
2. for cases that eof_handler not used(baselay),
no fetch eof would be generated
also workaround that keep fetch eof for HDMI as no HDMI converter eof
is kept.

Change-Id: I527b67339acfa08aa17da92754534a65130a3dab
Signed-off-by: Zhou Zhu <zzhu3@marvell.com>