pxa95xfb: reduce converter eof irq
reduce converter eof irq: it will be only unmasked when wait_for_vsync
and unmask when handled. an irq_pending counter is used to handle
multi wait_vsync calling case.
this mechanism is used for pxa978 only as for pxa955, a workaround
is required to do hss change in converter eof

Change-Id: Id65dab3e19ccc9a7c6229b49a953a62c200dbdaf
Signed-off-by: Zhou Zhu <zzhu3@marvell.com>