ov9740: Add software reset in stream-off sequence
when doing resolution switch from VGA to 720P, the sensor output
will be missing some lines sometime. In original open source
have a S/W reset in stream-off, we removed it because if reset is
not finished and we start the next set_fmt, sensor state-machine
maybe corrupted and cause IQ problem.

Now revert the change and add delay after reset to guarantee the
sensor is in good state.

1> Revert "ARM: ov9740: remove software reset in streamoff"
2> Add 5us delay after S/W reset

Change-Id: Id1c7868e7163eec178da74248939895f427280d8
Signed-off-by: Jiaquan Su <jqsu@marvell.com>