ARM: plat-pxa: add mfp lpm support for pxa988
In PXA988, the MFP_LPM_XXX configs would only take effect when the
chip enters chip level low power modes. The MFP_LPM_XXX controls
the behavior in LPM mode, if not set, the pin behaves the same as
in active mode. The MFP_PULL_XXX controls the pull setting for both
active and LPM modes. Select proper combinations to configure the
pins properly in pxa988 if necessary.

Example to set the pin output HIGH in lpm:
  MFP_CFG(GPIO22, AF1) | MFP_LPM_DRIVE_HIGH

Change-Id: I8a86d7d37fa03a83527029ee4e8995441e9fbc7b
Signed-off-by: Bin Wang <binw@marvell.com>