ARM: mmp: clean core morphing driver
Before we coupled core morphing code with cpufreq and hotplug, so that
core morphing driver will register the notifier with these two modules.
In current code base, we have refined the code to de-couple cpufreq and
hotplug, so need clean core morphing driver as well.

Now the core morphing's module is straightforward: if other module want
to constraint to only run on mp1, they it will vote with the function
*cm_vote_mp1()*; if the module don't have such constraint anymore, it
will use the function *cm_cancel_vote_mp1()*; so from the perspective of
core morphing, if there have no any vote for mp1 core, then it will swap
to mm core as possible so that can save the power.

Change-Id: Id87dacda49b04d721c8f28dcea294b57da7fdaec
Signed-off-by: Leo Yan <leoy@marvell.com>