ARM: pxa: power: set LPM single rail back after D1 exit
In previous implementation, we will set LPM using single rail when
entering D2, set LPM using dual rail when entering D1.

But whether Comm side is single rail or dual rail in its D2 is
determined by the setting when it is entering D2. And Comm side
should use single rail for better power consumption.

Using this patch, the dual rail setting can only appear when Apps
is in D1. So the possiblility that Comm side with dual rail setting
is less than previous implementation.

This patch can't totally avoid Comm side in D2 with dual rail
setting. But this is the best way we can do now.

Change-Id: If65b51af6fc799d476fade8b655ee1e23516a3b5
Signed-off-by: Lu Mao <ylmao@marvell.com>