ARM: mmp: add csi dphy rst operation for pxa988 ccic
Add csi dphy operation as it is missing in the register sheet.
Release it from reset prior to phy clk enable and reset it
after disable csi phy clock.
This is the reset control signal for CSI DPHY.
If DPHY is powered on, the calibration will be activated
once this signal is released.
At the initial stage, the calibration of CSI DPHY does not
work if it is in reset state.

Change-Id: Ibdc767cd27bd704d7ec33ea6b7f38bebb1dc2c1d
Signed-off-by: Zhoujie Wu <zjwu@marvell.com>