media: sci: move timestamp to EOF IRQ
According to V4L2 spec, TimeStamp should be set upon receive of
first byte of the frame, so we put TS in SOF IRQ. Unfortunatelly,
SCI will miss some SOF very occasionally, and this will cause a
wrong TS info (buffer returned with TS not updated, so a reverse
TS is ovserved).

Change TS to EOF IRQ as a W/R, and since we don't really care if
the TS is generated in SOF or EOF, this can be a solution.

Also add log to debug the SOF missing issue

Change-Id: I6c1d548e693adb5f8a757a5b8a4c667eccc8d79a
Signed-off-by: Jiaquan Su <jqsu@marvell.com>