ARM: pxa: power: don't change voltage if setrate when clock off
In previous code, there is a bug. Using current clock tree + dvfs,
if we want to setrate when clock is off, this component will start
request voltage when it is clock off.

To fix this, we will add check when setrate, if the clock is off,
we don't set voltage. So it is still relinquishing the voltage.

Change-Id: I2473f2b9f70d8c63f01f0025ec1025a7236ec5c0
Signed-off-by: Lu Mao <ylmao@marvell.com>