video: hdmi: enable continous 3d mode
CPU need to toggle left/right view in every vsync interrupt if 3D
is enabled. We switch to continous mode for one time mode can not
guarantee 3D formatter left/right view is aligned with frame_cnt[22].
frame_cnt[22:23] is video timing counter which is counting increasingly
as long as vs/hs/clk is on, even not DMA is active.

Change-Id: I7aa95265ecbebc7346183eb523e96885fe17d504
Signed-off-by: Jun Nie <njun@marvell.com>