ARM: pxa988: invalidate l1 dcache before flush
The secondary_startup will call v7_flush_dcache_all before it enables
the L1; however, the L1 comes out of reset in an undefined state, so the
clean and invalidate performed by v7_flush_dcache_all causes a bunch of
cache lines with uninitialized data and uninitialized tags to get
written out to memory, which does really unpleasant things to the main
processor. So we'd better use invalidate rather than a clean and
invalidate for secondary core, before jumping into the kernel.

Change-Id: I693ac058d45e2c308cc28de2286c462a7b00de51
Signed-off-by: Neil Zhang <zhangwm@marvell.com>
Signed-off-by: Bin Wang <binw@marvell.com>