ARM: pxa: set core voltage when cpufreq init
Request the voltage that core needs when cpufreq initialize
itself. this is used to prevent the protential issue:
OBM sets to 1.4G, this requires voltage level 3. but dvfs
initialized core voltage as level 2. and after cpufreq is
initialized, it uses performance governor to set cpu to max
1.4G. but dvfs found it is already 1.4G so there is no voltage
change. but the internal dvfs core voltage level is not right now.

Change-Id: I06caa77492e4c94e3370257a4c9d75db947f605d
Signed-off-by: Xiaoguang Chen <chenxg@marvell.com>