media: sci: fix memleak in vb2 context alloc
A memleak is detected in driver open, root caused as a vb2 context
alloced but not freed.

The context is allocated upon each time when driver is open, but
never release when close. Actually, the way is allocate it upon
vdev register, and release when vdev remove. Changed to the right
way.

Change-Id: Id4b58ee1fcc20ca954d19a81137f3611ca0e3a84
Signed-off-by: Jiaquan Su <jqsu@marvell.com>