hdmi: always disable clk in cable pull out
1. in clk_disable(), it will really turn off clk according to
counter, if calbe plug in and at this time delay_disable is
implemented, clk is not really disable, as clk_enable for 2 times
disable for 1 time, and when cable pull out, disable for 2 times,
really disable clk then. so, it should always clk_disalbe in
delay_disable(), otherwise, when next time cable pull out, clk
can not be disalbe when cable pull out.

2. reduce delay_disable() to 200ms

Change-Id: Ic9563e6fcb3f81ab11e2bdeb62abe0552ea384b2
Signed-off-by: Qing Xu <qingx@marvell.com>