ARM: mmp: add constraint for core morphing
Add CM's constraint for hotplug and cpufreq:

For hotplug, if the hotplug governor want to hotplug in mp2 core,
then it firstly need to vote mp1 core for CM, so that means there
must have mp1 and mp2 cores to run if system need two core to run;

For cpufreq, 400MHz is a threshold. If the core's frequency change
from <= 400MHz to > 400Mhz, then must set the constraint for CM so
that before DFC the kernel will firstly swap from mm core to MP1
core. The reason is we can get power benefit if mm core's frequency
is less than 400Mhz, so if core's frequency wants to change to high
value, then need change to mp1 core firstly.

NOTE: for cpufreq, here have a assumption: mm core has the same
frequency with mp1 core if the core frequency <= 400MHz; if so, we
can trigger core morphing seamlessly, because mp1 core will always
has the same frequency with mm core when the cores run with the
freq <= 400MHz; If the core run faster > 400MHz, we will set the
constraint for CM, so that mp1 core will be swapped in.

Change-Id: Id39b4e79e7fc0e13a84abbdfb4dd61e207e1513d
Signed-off-by: Leo Yan <leoy@marvell.com>