ARM: pxa: use C format IRQ handler to support unwind
When there is kernel panic or oops, the kernel will try to unwind the stack,
and will print all the functions that is stored in the stack. It is great
help to debug kernel bug with unwind information.

But previously, pxa95x_handle_irq_intc() was wroten in ASM, and didn't obey
the APCS (ARM Procedure Call Standard), so that this function cannot be
unwound. When unwind_backtrace() comes to this function, it will stop.
So functions in the stack before pxa95x_handle_irq_intc() can't be printed
out, we will lost many useful debug information.

I re-write this function in C format, so that it can be unwound, and functions
in the stack can be printed out. Please refer to arch/arm/kernel/unwind.c for
more information.

Change-Id: If49323333f925e669bacee9ef6e4af4d357a57a4
Signed-off-by: Leo Song <liangs@marvell.com>