ARM: pxa988: trigger IPC when sending IPI
When system enters M2 low power mode, the GIC loses all the setting.
If one core comes out from M2 due to proper ICU wake up source, and
tries to send IPI to wake up the other core, the other core can't be
waken up since the banked registers in the GIC for that core is not
restored yet.

This patch adds a workaround to handle this issue. When one core tries
sending IPI, it triggers IPC GP_INT interrupt as well. As the global
ICU interrupt is masked, the GP_INT interrupt only takes effect as
an wake up source to bring back the other core. The interrupt status
will be cleared before lpm.

Change-Id: I88e241383903e010f6c315e12c6111deff5d69f1
Signed-off-by: Bin Wang <binw@marvell.com>