ARM: mmp: Tune the L2 data / tag latency
The current data / tag latency is conservative (0x777) but safe. But the
performance is not so good.
The recommend latency for PL310 is 0x0 (L2 clk lower than 416MHz) and
0x010 (L2 clk higher than 416MHz). But it would be difficult to change
L2 latency on the fly. So just use the max value (0x010) before enable
L2 and never modify it dynamicly.

Change-Id: If6f2db8d9afd78cda50dabac27b621eaf5414ecd
Signed-off-by: Neil Zhang <zhangwm@marvell.com>