ARM: pxa: power: remove system wakeup wakelock
When system resume, we hold a 5s wakelock to make sure all the
wake up event can be handled by userspace.
But it is hard to decide the time, too long cause more power
consumption, too short may cause some event miss.

One way to resolve the problem is: hold the wakelock in driver,
unlock the wakelock for userspace after receiving the event.

The frist step is to remove the system wakeup wakelock.

Change-Id: I74378905134146027e7b49bc4b8d48a4b53b8d6a
Signed-off-by: Wenzeng Chen <wzch@marvell.com>