usb: gadget:fix USB controller reset repeatly.
The USB controller will reset repeatly while enumerated.
It exceeds 3ms when the device driver run to the point of
reading the portsc register. According to the USB spec.
This case is recognized as reset fail. So the driver
reset the USB contoller again. After debug, found that
the cause is that print operation(pr_info) in function
android_work. The pr_info occupied the CPU for a long time
resulting exceeding 3ms. So changed the pr_info to pr_debug
to save time.

Change-Id: I365245f2ddaa1ec6bf9a65e5fd39aafdfabc499c
Signed-off-by: carlpeng <carlpeng@marvell.com>