thermal: remove initial delay for pxa988
Background: Why initial delay first added ?
Uevent is a socket of type SOCK_DGRAM, that's mean, if the receiver
(Android service) is not ready when the sender(driver) sends the message,
the message is lost.

Remove: why it removed ?
The delay solution is not a good way for sync with driver and
Android service. Because it's hard to decide how long the delay
actually needed.

Solution:
When Android service startup, it need to read the temperature once and
do related cooling job, after that, Android service just wait for uevent

Change-Id: Id01806aeb2ee5a27854678a4c4bfb95d38a58f5a
Signed-off-by: hongfeng <hongfeng@marvell.com>