ARM: pxa988: switch to soc timer from twd in lpm
Local timer stops in the lpm mode when MP logic is powered off, while
the soc timer can still run and be able to generate interrupt and wake
up the cores. Now we make use both local timer and one soc timer to
solve this problem.

During cpu active and c1 mode, local timer is enabled, and works as
normal. The soc timer is registered as a broadcast timer, and generally
have no events set.

Before one tries to enter C2 or deeper, it calls notification of
CLOCK_EVT_NOTIFY_BROADCAST_ENTER, so that disable the local timer and
set the next event to the broadcast timer if necessary. Recover from
lpm accordingly.

When an soc timer interrupt comes, it checks the current time and the
"next" time for each core, if it expires, send IPI_TIMER to wake up
other cores. (No need IPI if it's for this core)

Change-Id: Ie300af4f890463612339584a11c4e0c9a2933513
Signed-off-by: Bin Wang <binw@marvell.com>