ARM: mmp: check pending interrupt befor wfi for pxa988
There is a corner case that Dragon will enter M2 even there is an
interrupt pending. The case can be discribed as following:
1. core 0 disable interrupt by CPSR.
2. core 1 send IPI to core 0
3. core 1 enter c2
4. core 0 mask GIC & ICU global mask
5. core 0 enter c2
There is an risk window that Dragon may enter M2 and powerdown before
PMU unmask GIC & ICU global mask and wakeup the core. It will cause the
pending IPI lost.
It will be fixed in Z3 and further, then we can discard this workaroud.

Change-Id: I5efa50268b0af92fea7acb96b018c1a2cb9bc747
Signed-off-by: Neil Zhang <zhangwm@marvell.com>