ARM: mmp: Do not resotre GIC when MP is not powered off
When reset handler finds the SCU is disabled, it means M2 was entered and MP
was powered down, so it will set a flag to notify low power mode function to
restore GIC distributor register context when it resumes back.

This can avoid the useless GIC restore when MP was actually not powered off.

Change-Id: I6b85a81edb262572f9006a2db1da0647f5ca356d
Signed-off-by: Raul Xiong <xjian@marvell.com>