ARM: mmp: emei_dkb: power off SD signal when suspend
There are two LDOs for SD card's power supply.
One is LDO13, which is for external card's VCC. Another is LDO12, which
is for SD bus signal. Card Detect pin also depond on LDO12.

At first, we keep LDO12 on and LDO13 off when suspend, but we find that
there is about 0.9mA leakage. So power off LDO12 in this patch for
suspend.
But you should know that CD pin doesn't work if LDO12 is off by current
HW design.

Change-Id: I2d8d0eb6021e5c7da6341728085b4cf85c4ecd6e
Signed-off-by: Jialing Fu <jlfu@marvell.com>