Do not use different ramdon MAC address on each ifup
LP: #1152330

When there's no valid content in EEPROM, a random MAC address will be
used. But if a different address is used on each ifup, network manager
will not do auto configuration on this interface.

Generate the random MAC address only once on driver bind and keeps using
this address until unbind to fix this issue.

Signed-off-by: Wen-chien Jesse Sung <jesse.sung@canonical.com>