dt-bindings: olpc,ap-sp: add GPIO lines
Add properties describing the GPIO lines used by the keyboard controller.
The olpc-apsp driver will do happily without them, but they are still part
of the hardware description. The driver could still reserve the lines,
so that nothing else touches them.

This makes the device node almost compatible with "ps2-gpio". I'm not adding
a compatible property, because ps2-gpio would use a different interrupt,
so that we'd need to name it. And I haven't actually tried it.

Signed-off-by: Lubomir Rintel <lkundrak@v3.sk>