ARM: mmp/mmp2: use cpu_is_pj4() instead of cpu_is_mmp2()
The MMP2 platform uses the PJ4 CPU. The cpu_is_mmp2() macro is thus actually
not useful at all and moreover gives the wrong result on MACH_MMP2_DT.

Thus, on a device-tree enabled system, the timer ends up being initialized
incorrectly. In fact, it ticks like at rate that's 1/100 slower or so.

Signed-off-by: Lubomir Rintel <lkundrak@v3.sk>