ARM: dts: add OLPC XO 1.75 machine
This is a fairly complete description of an OLPC XO 1.75 laptop.

It's notably missing:

* Vivante GPU.
* Audio.

I'll add those as time permits.

The laptop's native OpenFirmware is actually able to flatten the device
tree and pass it to the kernel. However, it generates faulty DTB and even
with the most recent firmware images, the differencies between the bindings
that firmware provides and the bindings that drivers actually expect are
just too significant.

Given the firmware is actually free software it might be possible to
fix it. Alternatively, it can just be patched around with a Forth boot
script.

Signed-off-by: Lubomir Rintel <lkundrak@v3.sk>