ARM: dts: add OLPC XO 1.75 machine
This is a fairly complete description of an OLPC XO 1.75 laptop. It's
missing a couple of GPIO lines for making the SDIO Wi-Fi actually work,
the GPIO lines to LID switches (there are two as the display can turn
around and be closed with screen facing up). I'll add those as time
permits.

The laptop's native OpenFirmware is actually able to flatten the device
tree and pass it to the kernel. However, even with the most recent
firmware images, there are significant differencies between the bindings
that firmware provides and the bindings that drivers actually expect.

Given the firmware is actually free software it might be possible to
fix it. Alternatively, it can just be patched around with a Forth boot
script.

So far, I've tested this with CONFIG_ARM_APPENDED_DTB.

Signed-off-by: Lubomir Rintel <lkundrak@v3.sk>